% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{citypop}
\alias{citypop}
\title{Most populous city in each state (2010)}
\format{A data frame with 51 rows and 5 variables.}
\usage{
data(citypop)
}
\description{
The most populous city in each US state, as of the 2010 US Census.\cr\cr
  The data is formatted for transforming with \code{\link[usmap_transform]{usmap_transform}}.
  Once the longitude and latitude is transformed, it can be added to
  \code{\link[plot_usmap]{plot_usmap}} using \code{ggplot2} layers.
}
\details{
\itemize{
  \item \code{lon} The longitude of the most populous city.
  \item \code{lat} The latitude of the most populous city.
  \item \code{state} The name of the state containing the city.
  \item \code{most_populous_city} The name of the city.
  \item \code{city_pop} The population of the city.
}
}
\references{
\itemize{
    \item \url{https://www.census.gov/programs-surveys/decennial-census/decade.2010.html}
  }
}
\keyword{data}
