% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.east_north_central}
\alias{.east_north_central}
\title{East North Central census division}
\format{An object of class \code{character} of length 5.}
\usage{
.east_north_central
}
\description{
US Census Bureau regional division containing Illinois, Indiana, Michigan,
Ohio, and Wisconsin.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .east_north_central, labels = TRUE)
}
\keyword{datasets}
