% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.mid_atlantic}
\alias{.mid_atlantic}
\title{Mid-Atlantic census division}
\format{An object of class \code{character} of length 3.}
\usage{
.mid_atlantic
}
\description{
US Census Bureau regional division containing New Jersey, New York,
and Pennsylvania.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .mid_atlantic, labels = TRUE)
}
\keyword{datasets}
