% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.mountain}
\alias{.mountain}
\title{Mountain census division}
\format{An object of class \code{character} of length 8.}
\usage{
.mountain
}
\description{
US Census Bureau regional division containing Arizona, Colorado, Idaho,
Montana, Nevada, New Mexico, Utah, and Wyoming.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .mountain, labels = TRUE)
}
\keyword{datasets}
