% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.new_england}
\alias{.new_england}
\title{New England census division}
\format{An object of class \code{character} of length 6.}
\usage{
.new_england
}
\description{
US Census Bureau regional division containing Connecticut, Maine,
Massachusetts, New Hampshire, Rhode Island, and Vermont.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .new_england, labels = TRUE)
}
\keyword{datasets}
