% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.north_central_region}
\alias{.north_central_region}
\title{North-Central census region}
\format{
An object of class \code{character} of length 12.
}
\usage{
.north_central_region
}
\description{
Former US Census Bureau region containing the East North Central and West
North Central divisions. This region has been designated as "Midwest"
since June 1984.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .north_central_region, labels = TRUE)
}
\keyword{datasets}
