% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.northeast_region}
\alias{.northeast_region}
\title{Northeast census region}
\format{An object of class \code{character} of length 9.}
\usage{
.northeast_region
}
\description{
US Census Bureau region containing the New England and Mid-Atlantic
divisions.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .northeast_region, labels = TRUE)
}
\keyword{datasets}
