% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.pacific}
\alias{.pacific}
\title{Pacific census division}
\format{An object of class \code{character} of length 5.}
\usage{
.pacific
}
\description{
US Census Bureau regional division containing Alaska, California, Hawaii,
Oregon, and Washington.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .pacific, labels = TRUE)
}
\keyword{datasets}
