% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.south_atlantic}
\alias{.south_atlantic}
\title{South Atlantic census division}
\format{An object of class \code{character} of length 9.}
\usage{
.south_atlantic
}
\description{
US Census Bureau regional division containing Delaware, Florida, Georgia,
Maryland, North Carolina, South Carolina, Virginia,
District of Columbia, and West Virginia.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .south_atlantic, labels = TRUE)
}
\keyword{datasets}
