% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.south_region}
\alias{.south_region}
\title{South census region}
\format{An object of class \code{character} of length 17.}
\usage{
.south_region
}
\description{
US Census Bureau region containing the South Atlantic, East South Central,
and West South Central divisions.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .midwest_region, labels = TRUE)
}
\keyword{datasets}
