% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.west_north_central}
\alias{.west_north_central}
\title{West North Central census division}
\format{
An object of class \code{character} of length 7.
}
\usage{
.west_north_central
}
\description{
US Census Bureau regional division containing Iowa, Kansas, Minnesota,
Missouri, Nebraska, North Dakota, and South Dakota.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .west_north_central, labels = TRUE)
}
\keyword{datasets}
