% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.west_region}
\alias{.west_region}
\title{West census region}
\format{
An object of class \code{character} of length 13.
}
\usage{
.west_region
}
\description{
US Census Bureau region containing the Mountain and Pacific divisions.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .midwest_region, labels = TRUE)
}
\keyword{datasets}
