% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{.west_south_central}
\alias{.west_south_central}
\title{West South Central census division}
\format{An object of class \code{character} of length 4.}
\usage{
.west_south_central
}
\description{
US Census Bureau regional division containing Arkansas, Louisiana, Oklahoma,
and Texas.
}
\details{
See
\url{https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf}
}
\examples{
plot_usmap(include = .west_south_central, labels = TRUE)
}
\keyword{datasets}
