% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{fips}
\alias{fips}
\title{Retrieve FIPS code for either a US state or county}
\usage{
fips(state = "", county = "")
}
\arguments{
\item{state}{The state for which to obtain a FIPS code.}

\item{county}{The county for which to obtain a FIPS code.}
}
\value{
The FIPS code of given `state` or `county`.
}
\description{
Retrieve FIPS code for either a US state or county
}
\note{
Either a `state` or a `county` should be specified, but not both.
}
\examples{
print("Examples will go here.")
}

