% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{get_fips_info}
\alias{get_fips_info}
\title{Gets FIPS info for either states or counties depending on input.
Helper function for S3 method \code{fips_info}.}
\usage{
get_fips_info(fips, sortAndRemoveDuplicates)
}
\description{
Gets FIPS info for either states or counties depending on input.
Helper function for S3 method \code{fips_info}.
}
\keyword{internal}
