% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{getFipsInfo}
\alias{getFipsInfo}
\title{Gets FIPS info for either states or counties depending on input.
Helper function for S3 method \code{fips_info}.}
\usage{
getFipsInfo(fips)
}
\description{
Gets FIPS info for either states or counties depending on input.
Helper function for S3 method \code{fips_info}.
}
\keyword{internal}

