% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{perform_transform}
\alias{perform_transform}
\title{Transform \code{sf} coordinates to \code{usmap} transform}
\usage{
perform_transform(data, ...)
}
\description{
Internal function with common functionality for transforming coordinates.
Using this function directly is not recommended.
}
\keyword{internal}
