% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{plot_usmap}
\alias{plot_usmap}
\title{Conveniently plot basic US map}
\usage{
plot_usmap(regions = "states", include = c(), theme = theme_map())
}
\arguments{
\item{regions}{The region breakdown for the map, either \code{"states"}
or \code{"counties"}.}
  
\item{include}{The region (state names or state/county FIPS codes) to include
in the map. Usage of this parameter is identical to the \code{\link[usmap]{us_map}} 
function.}

\item{theme}{The theme that should be used for plotting the map. The default
is \code{\link[ggthemes]{theme_map}}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that contains a basic
  US map with the described parameters. If \code{ggplot2} is not installed,
  \code{\link[graphics]{plot}} is used, which may result in slower execution.
  Moreover, basic plots cannot be stored in a variable or customized (themes, scales, etc.)
  like \code{ggplot} can so it is highly recommend that \code{ggplot2} be installed
  for a much better plotting experience.
}
\description{
\code{plot_usmap} quickly and easily plots a blank US map
  using \pkg{ggplot2} if it is installed.
}
\examples{
plot_usmap()
plot_usmap(regions = "states")
plot_usmap(regions = "counties")
plot_usmap(regions = "state")
plot_usmap(regions = "county")
  
# Output is ggplot object so it can be extended
#  with any number of ggplot layers
library(ggplot2)
plot_usmap(regions = "states", include = c("CA", "NV", "ID", "OR", "WA")) +
  labs(title = "Western States")

}
\seealso{
\code{\link{usmap}}
}
