% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{set_sp_evolution_status}
\alias{set_sp_evolution_status}
\title{Set sp evolution status}
\usage{
set_sp_evolution_status()
}
\description{
Sets the \code{sp} evolution status to "2" to
force usage of \code{sf} instead of \code{rgdal}
which is being retired.

This can be removed in the future when the evolution status
is set to >= 2 by default in \code{sf}.
}
\keyword{internal}
