% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{statepop}
\alias{statepop}
\title{Population estimates (2015), state level}
\format{
A data frame with 51 rows and 4 variables.
}
\usage{
data(statepop)
}
\description{
US census population estimates by state for 2015. \cr\cr
  The data is formatted for easy merging with output from \code{\link[usmap]{us_map}}.
}
\details{
\itemize{
  \item \code{fips} The 2-digit FIPS code corresponding to the state.
  \item \code{abbr} The 2-letter state abbreviation.
  \item \code{full} The full state name.
  \item \code{pop_2015} The 2015 population estimate (in number of people)
    for the corresponding state.
}
}
\references{
\itemize{
    \item \url{https://www.census.gov/programs-surveys/popest.html}
    \item \url{https://www.ers.usda.gov/data-products/county-level-data-sets/}
  }
}
\keyword{data}
