% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{static_merge}
\alias{static_merge}
\title{Performs merge while maintaining original sort order.}
\usage{
static_merge(x, y, ...)
}
\description{
Performs merge while maintaining original sort order.
}
\seealso{
https://stackoverflow.com/a/61560405/7264964
}
\keyword{internal}
