% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_sf}
\alias{transform_sf}
\title{Transform \code{sf} coordinates to \code{usmap} transform}
\usage{
transform_sf(data_sf)
}
\description{
Internal function with common functionality for transforming coordinates.
Using this function directly is not recommended.
}
\keyword{internal}
