% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usmap.R
\name{us_map}
\alias{us_map}
\title{Retrieve US map data}
\usage{
us_map(regions, include = c())
}
\arguments{
\item{regions}{The region breakdown for the map, either (\code{"states"})` 
or (\code{"counties"}).}

\item{include}{The regions to include in the output data frame. The regions
must be of the same type as \code{regions}.}
}
\value{
A data frame of US map coordinates divided by the desired `regions`.
}
\description{
Retrieve US map data
}
\examples{
df <- us_map(regions = "states")
plot(df$long, df$lat)
}

