% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{usmap_transform}
\alias{usmap_transform}
\alias{usmap_transform.sf}
\alias{usmap_transform.data.frame}
\title{Convert coordinate date frame to usmap projection}
\usage{
usmap_transform(data, input_names, output_names)

\method{usmap_transform}{sf}(data)

\method{usmap_transform}{data.frame}(data, input_names = c("lon", "lat"), output_names = NULL)
}
\arguments{
\item{data}{A data frame containing coordinates in a two column format
where the first column represents longitude and the second data frame
represents latitude. The names of the data frame column do not matter,
just that the order of the columns is kept intact.}

\item{input_names}{A character vector of length two which specifies the
longitude and latitude columns of the input data (the ones that should be
transformed), respectively. Not required if the input data is
an \code{sf} object. Defaults to \code{c("lon", "lat")}.}

\item{output_names}{Defunct, this parameter is no longer used. The output
of this function will have a column named \code{"geometry"} with the transformed
coordinates.}
}
\value{
An \code{sf} object containing the transformed coordinates from the
input data frame with the US National Atlas Equal Area projection applied.
The transformed columns will be appended to the data frame so that all
original columns should remain intact.
}
\description{
Converting an external data frame of map coordinates will
allow those points to line up with the regular usmap plot by applying
the same US National Atlas Equal Area projection to those points as well.
}
\examples{
data <- data.frame(
  lon = c(-74.01, -95.36, -118.24, -87.65, -134.42, -157.86),
  lat = c(40.71, 29.76, 34.05, 41.85, 58.30, 21.31),
  pop = c(8398748, 2325502, 3990456, 2705994, 32113, 347397)
)

# Transform data
transformed_data <- usmap_transform(data)

# Plot transformed data on map
library(ggplot2)

plot_usmap() + geom_sf(
  data = transformed_data,
  aes(size = pop),
  color = "red", alpha = 0.5
)

}
