% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{countypop}
\alias{countypop}
\title{Population estimates (2015), county level}
\format{
A data frame with 3142 rows and 4 variables.
}
\usage{
data(countypop)
}
\description{
US census population estimates by county for 2015. \cr\cr
The data is formatted for easy merging with output from \link{us_map}.
}
\details{
\itemize{
\item \code{fips} The 5-digit FIPS code corresponding to the county.
\item \code{abbr} The 2-letter state abbreviation.
\item \code{county} The full county name.
\item \code{pop_2015} The 2015 population estimate (in number of people)
for the corresponding county.
}
}
\references{
\itemize{
\item \url{https://www.census.gov/programs-surveys/popest.html}
\item \url{https://www.ers.usda.gov/data-products/county-level-data-sets/}
}
}
\keyword{data}
