% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{fips}
\alias{fips}
\title{Retrieve FIPS code for either a US state or county}
\usage{
fips(state, county = "")
}
\arguments{
\item{state}{The state for which to obtain a FIPS code.
Can be entered as either a state abbrevation or full name (case-insensitive).}

\item{county}{The county for which to obtain a FIPS code.
Can be entered with or without "county" (case-insensitive).}
}
\value{
The FIPS code of given \code{state} or \code{county}.
}
\description{
Retrieve FIPS code for either a US state or county
}
\note{
A `state` must be included when searching for `county`,
 otherwise multiple results may be returned for duplicate county names.
}
\examples{
fips("NJ")
fips("California")
fips("CA", county = "orange")
fips(state = "AL", county = "autauga")
fips(state = "Alabama", county = "Autauga County")
}

