% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{theme_map}
\alias{theme_map}
\title{Convenient theme map}
\usage{
theme_map(base_size = 9, base_family = "")
}
\description{
This creates a nice map theme for use in \link{plot_usmap}.
It is borrowed from the \code{ggthemes} package located at this repository:
https://github.com/jrnold/ggthemes.

This function was manually rewritten here to avoid the need for
another package import.

All theme functions (i.e. \code{theme_bw}, \code{theme}, \code{element_blank}, \verb{\%+replace\%})
come from \code{ggplot2}.
}
\keyword{internal}
