% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usmap.R
\name{us_map}
\alias{us_map}
\title{Retrieve US map data}
\usage{
us_map(region, include = c())
}
\arguments{
\item{region}{The region breakdown for the map, either `states` or `counties`.}

\item{include}{The regions to include in the output data frame. The regions
must be of the same type as `region`.}
}
\value{
A data frame of US map coordinates divided by the desired `region`.
}
\description{
Retrieve US map data
}
\examples{
library(usmap)
states_map <- us_map(region = "states")
counties_map <- us_map(region = "counties")
}

