% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usmap-package.R
\docType{package}
\name{usmap}
\alias{usmap-package}
\alias{usmap}
\title{usmap: US maps including Alaska and Hawaii}
\description{
It is usually difficult or inconvenient to create US maps that
include Alaska, Hawaii, and even Puerto Rico in a convenient spot. All map
data frames produced by this package use the US National Atlas Equal Area
projection.
}
\section{Map data}{

Alaska, Hawaii, and Puerto Rico have been manually moved to a new location so that
their new coordinates place them to the bottom-left corner of
the map. These maps can be accessed by using the \code{\link[=us_map]{us_map()}} function.

The function provides the ability to retrieve maps with either
state borders or county borders using the \code{regions} parameter
for convenience.

States (or counties) can be included and excluded using the provided
\code{include} and \code{exclude} parameters. These parameters can be used
together with any combination of names, abbreviations, or FIPS code to
create more complex maps.
}

\section{FIPS lookup tools}{

Several functions have been included to lookup the US state or county
pertaining to a FIPS code.

Likewise a reverse lookup can be done where a FIPS code can be used to
retrieve the associated states or counties. This can be useful when
preparing data to be merged with the map data frame.
}

\section{Plot US map data}{

A convenience function \code{\link[=plot_usmap]{plot_usmap()}} has been included which
takes similar parameters to \code{\link[=us_map]{us_map()}} and returns a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2}
object. Since the output is a \code{ggplot} object, other layers can be
added such as scales, themes, and labels. Including data in the function call
will color the map according to the values in the data, creating a choropleth.
}

\section{Transforming data}{

It is also possible to add spatial data to the map, in the form of either
data frames or simple features (\link[sf:sf]{sf::sf}) objects. If necessary, the
data can be transformed to be in the same coordinate reference system as
\link{usmap} by using \code{\link[=usmap_transform]{usmap_transform()}} and then plotted using \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.
}

\references{
Rudis, Bob. "Moving The Earth (well, Alaska & Hawaii) With R."
Blog post. Rud.is., 16 Nov. 2014. Web. 10 Aug. 2015.
\url{https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/}.
}
\seealso{
Helpful links:
\itemize{
\item FIPS code information \cr
\url{https://en.wikipedia.org/wiki/FIPS_county_code}
\url{https://en.wikipedia.org/wiki/FIPS_state_code}
\item US Census Shapefiles \cr
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/cartographic-boundary.html}
\item Map Features \cr
\url{https://en.wikipedia.org/wiki/Map_projection} \cr
\url{https://en.wikipedia.org/wiki/Equal-area_projection} \cr
\url{https://en.wikipedia.org/wiki/Choropleth} \cr
\url{https://epsg.io/9311} (US National Atlas Equal Area)
}
}
\author{
Paolo Di Lorenzo \cr
\itemize{
\item Email: \email{dilorenzo@hey}
\item GitHub: \url{https://github.com/pdil/}
}
}
