% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usmap.R
\docType{package}
\name{usmap}
\alias{usmap}
\alias{usmap-package}
\title{usmap: US maps including Alaska and Hawaii}
\description{
It is usually difficult or inconvenient to create US maps that
  include both Alaska and Hawaii in a convenient spot.
}
\section{Map data frames}{

Alaska and Hawaii have been manually moved to a new location so that
their new coordinates place them to the bottom-left corner of
the map. These maps can be accessed by using the \code{\link{usmap}} function.

The function provides the ability to retrieve maps with either
state borders or county borders using the \code{regions} parameter
for convenience.

States (or counties) can be included such that all other states (or counties)
are excluded using the \code{include} parameter.
}

\section{FIPS lookup tools}{

Several functions have been included to lookup the US state or county
pertaining to a FIPS code.

Likewise a reverse lookup can be done where a FIPS code can be used to
retrieve the associated state(s) or county(ies). This can be useful when
preparing data to be merged with the map data frame.
}

\section{Merge data with map}{

Before plotting, the data can be merged with the map data frame internally
and returned as a complete data frame, ready for use. This is done by using
a data frame in which each row is identified by a state, county, or FIPS
code (and specifying which type). The returned data frame can be easily
plotted with \code{\link[graphics]{plot}} or \pkg{ggplot2} (or any other plotting solution).
}

