% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\docType{data}
\name{usmap_crs}
\alias{usmap_crs}
\title{usmap coordinate reference system}
\format{An object of class \code{CRS} of length 1.}
\usage{
usmap_crs
}
\description{
This coordinate reference system (CRS) represents
the canonical projection used by the \code{usmap} package. It can
be used to transform shape files, spatial points, spatial data
frames, etc. to the same coordinate representation that is used
by the \code{plot_usmap} function.
}
\keyword{datasets}
